src/requirements.txt
torch
pillow
scikit-learn
geopandas
matplotlib
numpy
osmnx
shapely
Pillow
srai[osm]
pyyaml
geopy
